Tidy up imports to be consistent with relative usage.
This change fixes an issue where on my desktop the create test that
raises an exception where a tsdb already exists was failing.

Not sure why it started happening now, but importing the exception in
different ways was confusing the unit test code.

Destkop being 64 bit Fedora 19 with a Python 3.3.2 virtual env.