Write method now takes pandas.Series instead of tuple
THIS IS A BREAKING CHANGE TO THE DATABASE API.

Since Pandas is a dependency already and it is great for reading many
formats of time series data the write method has been changed to take a
pandas Series object instead of the previous tuple of arrays.

Used vim macros in earnest for the first time to much success. Which
made a long and tedious change not take all that long. Meaning there is
still quite a bit more bus ride to go... Not even at Ballarat yet.

Signed-off-by: Andrew MacDonald <andrew@maccas.net>