Do a joined query to improve performance
When running db.list_timeseries_instances with a database containing a
large number of series instances performance can be slow due to lazy
loading running an individual query for each record when building the
data frame. Doing a join means all the records are read in a single
query, which is much more efficient when we are going to use all the
returned values anyway.

Signed-off-by: Andrew MacDonald <andrew@maccas.net>