Revert "Add indexes to meta-database"
This reverts commit 3b2a2390d91631ca31dfb0075a2a1fe8a2df9cd3.

This change is working towards a new format version, but as I haven't
made progress on that yet reverting this commit to enable merging of dev
branch with v0.6 and continuing bug fixes here. This commit will remain
in the v0.7 branch which will be used for progressing the core format
change work.