Switch to using a conda env specifically in Travis
Turns out pysocks is required in the root conda environment, in
hindsight that makes sense why it was there but I couldn't see what was
triggering the install in a local conda env.

New strategy, create a conda environment with the required version of
Python for testing, rather than just mashing everything into the root
environment. Maybe this will finally sort out the Python 3.4 build woes.