Don't test read_dataframe based on column order
The test failed in a Travis CI run because the order of the columns had
changed. Some combinations of Python 3.6.6 and Pandas 0.23.1 resulted in
the order being different. We don't require the data being in a specific
order so just test that all the expected columns were included instead.

The failing Travis CI job in question:
    https://travis-ci.org/amacd31/phildb/jobs/400194918

Signed-off-by: Andrew MacDonald <andrew@maccas.net>